std: move `errno` and related functions into `sys::io`
