+++ library/std/src/sys/pal/hermit/os.rs
@@ -5,14 +5,6 @@ use crate::path::{self, PathBuf};
 use crate::sys::unsupported;
 use crate::{fmt, io};
 
-pub fn errno() -> i32 {
-    unsafe { hermit_abi::get_errno() }
-}
-
-pub fn error_string(errno: i32) -> String {
-    hermit_abi::error_string(errno).to_string()
-}
-
 pub fn getcwd() -> io::Result<PathBuf> {
     Ok(PathBuf::from("/"))
 }
