+++ library/std/src/sys/pal/motor/mod.rs
@@ -44,45 +44,6 @@ pub fn unsupported_err() -> io::Error {
     io::Error::UNSUPPORTED_PLATFORM
 }
 
-pub fn is_interrupted(_code: io::RawOsError) -> bool {
-    false // Motor OS doesn't have signals.
-}
-
-pub fn decode_error_kind(code: io::RawOsError) -> io::ErrorKind {
-    use moto_rt::error::*;
-
-    if code < 0 || code > u16::MAX.into() {
-        return io::ErrorKind::Uncategorized;
-    }
-
-    let error = moto_rt::Error::from(code as moto_rt::ErrorCode);
-
-    match error {
-        moto_rt::Error::Unspecified => io::ErrorKind::Uncategorized,
-        moto_rt::Error::Unknown => io::ErrorKind::Uncategorized,
-        moto_rt::Error::NotReady => io::ErrorKind::WouldBlock,
-        moto_rt::Error::NotImplemented => io::ErrorKind::Unsupported,
-        moto_rt::Error::VersionTooHigh => io::ErrorKind::Unsupported,
-        moto_rt::Error::VersionTooLow => io::ErrorKind::Unsupported,
-        moto_rt::Error::InvalidArgument => io::ErrorKind::InvalidInput,
-        moto_rt::Error::OutOfMemory => io::ErrorKind::OutOfMemory,
-        moto_rt::Error::NotAllowed => io::ErrorKind::PermissionDenied,
-        moto_rt::Error::NotFound => io::ErrorKind::NotFound,
-        moto_rt::Error::InternalError => io::ErrorKind::Other,
-        moto_rt::Error::TimedOut => io::ErrorKind::TimedOut,
-        moto_rt::Error::AlreadyInUse => io::ErrorKind::AlreadyExists,
-        moto_rt::Error::UnexpectedEof => io::ErrorKind::UnexpectedEof,
-        moto_rt::Error::InvalidFilename => io::ErrorKind::InvalidFilename,
-        moto_rt::Error::NotADirectory => io::ErrorKind::NotADirectory,
-        moto_rt::Error::BadHandle => io::ErrorKind::InvalidInput,
-        moto_rt::Error::FileTooLarge => io::ErrorKind::FileTooLarge,
-        moto_rt::Error::NotConnected => io::ErrorKind::NotConnected,
-        moto_rt::Error::StorageFull => io::ErrorKind::StorageFull,
-        moto_rt::Error::InvalidData => io::ErrorKind::InvalidData,
-        _ => io::ErrorKind::Uncategorized,
-    }
-}
-
 pub fn abort_internal() -> ! {
     core::intrinsics::abort();
 }
