+++ library/std/src/sys/pal/solid/os.rs
@@ -1,4 +1,4 @@
-use super::{error, itron, unsupported};
+use super::{itron, unsupported};
 use crate::ffi::{OsStr, OsString};
 use crate::path::{self, PathBuf};
 use crate::{fmt, io};
@@ -9,14 +9,6 @@ impl itron::error::ItronError {
     pub(crate) fn as_io_error(self) -> crate::io::Error {
         crate::io::Error::from_raw_os_error(self.as_raw())
     }
-}
-
-pub fn errno() -> i32 {
-    0
-}
-
-pub fn error_string(errno: i32) -> String {
-    if let Some(name) = error::error_name(errno) { name.to_owned() } else { format!("{errno}") }
 }
 
 pub fn getcwd() -> io::Result<PathBuf> {
