+++ library/std/src/sys/pal/teeos/os.rs
@@ -7,10 +7,6 @@ use crate::ffi::{OsStr, OsString};
 use crate::path::PathBuf;
 use crate::{fmt, io, path};
 
-pub fn errno() -> i32 {
-    unsafe { (*libc::__errno_location()) as i32 }
-}
-
 // Hardcoded to return 4096, since `sysconf` is only implemented as a stub.
 pub fn page_size() -> usize {
     // unsafe { libc::sysconf(libc::_SC_PAGESIZE) as usize };
@@ -18,10 +14,6 @@ pub fn page_size() -> usize {
 }
 
 // Everything below are stubs and copied from unsupported.rs
-
-pub fn error_string(_errno: i32) -> String {
-    "error string unimplemented".to_string()
-}
 
 pub fn getcwd() -> io::Result<PathBuf> {
     unsupported()
