+++ library/std/src/sys/pal/unix/os.rs
@@ -14,134 +14,7 @@ use crate::sys::cvt;
 use crate::sys::helpers::run_path_with_cstr;
 use crate::{fmt, io, iter, mem, ptr, slice, str};
 
-const TMPBUF_SZ: usize = 128;
-
 const PATH_SEPARATOR: u8 = b':';
-
-unsafe extern "C" {
-    #[cfg(not(any(target_os = "dragonfly", target_os = "vxworks", target_os = "rtems")))]
-    #[cfg_attr(
-        any(
-            target_os = "linux",
-            target_os = "emscripten",
-            target_os = "fuchsia",
-            target_os = "l4re",
-            target_os = "hurd",
-        ),
-        link_name = "__errno_location"
-    )]
-    #[cfg_attr(
-        any(
-            target_os = "netbsd",
-            target_os = "openbsd",
-            target_os = "cygwin",
-            target_os = "android",
-            target_os = "redox",
-            target_os = "nuttx",
-            target_env = "newlib"
-        ),
-        link_name = "__errno"
-    )]
-    #[cfg_attr(any(target_os = "solaris", target_os = "illumos"), link_name = "___errno")]
-    #[cfg_attr(target_os = "nto", link_name = "__get_errno_ptr")]
-    #[cfg_attr(any(target_os = "freebsd", target_vendor = "apple"), link_name = "__error")]
-    #[cfg_attr(target_os = "haiku", link_name = "_errnop")]
-    #[cfg_attr(target_os = "aix", link_name = "_Errno")]
-    // SAFETY: this will always return the same pointer on a given thread.
-    #[unsafe(ffi_const)]
-    pub safe fn errno_location() -> *mut c_int;
-}
-
-/// Returns the platform-specific value of errno
-#[cfg(not(any(target_os = "dragonfly", target_os = "vxworks", target_os = "rtems")))]
-#[inline]
-pub fn errno() -> i32 {
-    unsafe { (*errno_location()) as i32 }
-}
-
-/// Sets the platform-specific value of errno
-// needed for readdir and syscall!
-#[cfg(all(not(target_os = "dragonfly"), not(target_os = "vxworks"), not(target_os = "rtems")))]
-#[allow(dead_code)] // but not all target cfgs actually end up using it
-#[inline]
-pub fn set_errno(e: i32) {
-    unsafe { *errno_location() = e as c_int }
-}
-
-#[cfg(target_os = "vxworks")]
-#[inline]
-pub fn errno() -> i32 {
-    unsafe { libc::errnoGet() }
-}
-
-#[cfg(target_os = "rtems")]
-#[inline]
-pub fn errno() -> i32 {
-    unsafe extern "C" {
-        #[thread_local]
-        static _tls_errno: c_int;
-    }
-
-    unsafe { _tls_errno as i32 }
-}
-
-#[cfg(target_os = "dragonfly")]
-#[inline]
-pub fn errno() -> i32 {
-    unsafe extern "C" {
-        #[thread_local]
-        static errno: c_int;
-    }
-
-    unsafe { errno as i32 }
-}
-
-#[cfg(target_os = "dragonfly")]
-#[allow(dead_code)]
-#[inline]
-pub fn set_errno(e: i32) {
-    unsafe extern "C" {
-        #[thread_local]
-        static mut errno: c_int;
-    }
-
-    unsafe {
-        errno = e;
-    }
-}
-
-/// Gets a detailed string description for the given error number.
-pub fn error_string(errno: i32) -> String {
-    unsafe extern "C" {
-        #[cfg_attr(
-            all(
-                any(
-                    target_os = "linux",
-                    target_os = "hurd",
-                    target_env = "newlib",
-                    target_os = "cygwin"
-                ),
-                not(target_env = "ohos")
-            ),
-            link_name = "__xpg_strerror_r"
-        )]
-        fn strerror_r(errnum: c_int, buf: *mut c_char, buflen: libc::size_t) -> c_int;
-    }
-
-    let mut buf = [0 as c_char; TMPBUF_SZ];
-
-    let p = buf.as_mut_ptr();
-    unsafe {
-        if strerror_r(errno as c_int, p, buf.len()) < 0 {
-            panic!("strerror_r failure");
-        }
-
-        let p = p as *const _;
-        // We can't always expect a UTF-8 environment. When we don't get that luxury,
-        // it's better to give a low-quality error message than none at all.
-        String::from_utf8_lossy(CStr::from_ptr(p).to_bytes()).into()
-    }
-}
 
 #[cfg(target_os = "espidf")]
 pub fn getcwd() -> io::Result<PathBuf> {
